adds responses

+++ src/main/java/com/veterinary/practica/domains/repositories/clients/SocialNetworksClientsRepository.java
@@ -1,7 +1,7 @@
 package com.veterinary.practica.domains.repositories.clients;
 
-import com.veterinary.practica.domains.entities.clients.SocialNetworksClientsEntity;
+import com.veterinary.practica.api.models.responses.clients.SocialNetworksClientsEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SocialNetworksClientsRepository extends CrudRepository<SocialNetworksClientsEntity, Long> {
+public interface SocialNetworksClientsRepository extends CrudRepository<SocialNetworksClientsEntity, Integer> {
 }
